backend-1/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-cosmos
azure-identity
Werkzeug
uuid7
requests
azure.storage.blob 
azure.identity
google-genai
Pillow
pytz